+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to this project will be documented in this file.
 - Bump Ruby 2.6.6 to 2.7.1 [#985](https://github.com/sider/runners/pull/985)
 - [Reek] Update reek requirement from 5.6.0 to 6.0.0 [#919](https://github.com/sider/runners/pull/919)
 - [Checkstyle] Bump checkstyle from 8.30 to 8.31 [#920](https://github.com/sider/runners/pull/920)
+- [Checkstyle] Bump checkstyle from 8.31 to 8.32 [#1028](https://github.com/sider/runners/pull/1028)
 - [FxCop] [roslyn-analyzers-runner](https://github.com/sider/roslyn-analyzers-runner) provides a static code analysis without running build [#971](https://github.com/sider/runners/pull/971)
 - [PMD Java] Bump pmd-java from 6.21.0 to 6.22.0 [#922](https://github.com/sider/runners/pull/922)
 - [PMD Java] Bump pmd-java from 6.22.0 to 6.23.0 [#1029](https://github.com/sider/runners/pull/1029)
@@ -27,6 +28,8 @@ All notable changes to this project will be documented in this file.
 - [GolangCI-Lint] Bump golangci/golangci-lint from v1.24.0 to v1.25.1 [#1023](https://github.com/sider/runners/pull/1023)
 - [FxCop] Bump FxCop from 2.9.8 to 3.0.0 [#1020](https://github.com/sider/runners/pull/1020)
 - [hadolint] Bump hadolint/hadolint from v1.17.5-debian to v1.17.6-debian [#1025](https://github.com/sider/runners/pull/1025)
+- [PHP_CodeSniffer] Bump squizlabs/php_codesniffer from 3.5.4 to 3.5.5 [#1032](https://github.com/sider/runners/pull/1032)
+- [Checkstyle] Shorten official issue ID [#1035](https://github.com/sider/runners/pull/1035)
 
 ## 0.22.4